Split up WINTER generator
