+++ mirar/utils/__init__.py
@@ -9,3 +9,4 @@ from mirar.utils.execute_cmd import (
     run_docker,
     run_local,
 )
+from mirar.utils.ldac_tools import get_table_from_ldac
